Fix event source corruption issue caused by HttpStatusCode enum.

+++ src/WebJobs.Extensions.DurableTask/EndToEndTraceHelper.cs
@@ -280,7 +280,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
                 version,
                 instanceId,
                 details,
-                statusCode,
+                (int)statusCode,
                 reason,
                 functionType,
                 ExtensionVersion,
@@ -289,7 +289,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
 
             this.logger.LogInformation(
                 "{instanceId}: Function '{functionName} ({functionType})', sent a '{functionState}' notification event to Azure Event Grid. Status code: {statusCode}. Details: {details}. HubName: {hubName}. AppName: {appName}. SlotName: {slotName}. ExtensionVersion: {extensionVersion}. SequenceNumber: {sequenceNumber}. Latency: {latencyMs} ms.",
-                instanceId, functionName, functionType, functionState, statusCode, details, hubName, LocalAppName, LocalSlotName, ExtensionVersion, this.sequenceNumber++, latencyMs);
+                instanceId, functionName, functionType, functionState, (int)statusCode, details, hubName, LocalAppName, LocalSlotName, ExtensionVersion, this.sequenceNumber++, latencyMs);
         }
 
         public void EventGridFailed(
@@ -315,7 +315,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
                 version,
                 instanceId,
                 details,
-                statusCode,
+                (int)statusCode,
                 reason,
                 functionType,
                 ExtensionVersion,
@@ -324,7 +324,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
 
             this.logger.LogError(
                 "{instanceId}: Function '{functionName} ({functionType})', failed to send a '{functionState}' notification event to Azure Event Grid. Status code: {statusCode}. Details: {details}. HubName: {hubName}. AppName: {appName}. SlotName: {slotName}. ExtensionVersion: {extensionVersion}. SequenceNumber: {sequenceNumber}. Latency: {latencyMs} ms.",
-                instanceId, functionName, functionType, functionState, statusCode, details, hubName, LocalAppName, LocalSlotName, ExtensionVersion, this.sequenceNumber++, latencyMs);
+                instanceId, functionName, functionType, functionState, (int)statusCode, details, hubName, LocalAppName, LocalSlotName, ExtensionVersion, this.sequenceNumber++, latencyMs);
         }
 
         public void TimerExpired(
